Add save conversion for Jovan quest (#181)
There's a very small edge case remaining where if a player saves on the beach, hasn't beaten the Avatar of Vigoroth yet, then does and wipes back to the Pokemon Centre, they can skip Jovan, but I'm not sure of a better flag to check for this scenario, and hopefully this should be a very rare situation if it comes up at all. Most players will either be before the beach (and so see Jovan naturally), after the avatar (and so have the event cleared for them), or survive the trip back from the avatar to the other side of the beach (and so run into Jovan).