eliminating instances of "multi"
made a few changes, mostly to things that reference spread moves, to eliminate the key phrase "multi". this is intended to make a text search in the movedex/masterdex ability text search for "multi" return exactly multi-hit moves. i believe this also standardizes the use of the term "spread move" in the process

anyways feel free to reject this change as it's a bit out of scope, left it in a separate commit so it could be an easy yes/no