Implement new Kanto Legendary Birds abilities
Galvanic Wings: Foes that become numbed are severely so. The Pokémon's Flying-type moves have a 30% chance to severely numb.
Glacial Wings: Foes that become frostbitten are severely so. The Pokémon's Flying-type moves have a 30% chance to severely frostbite.
Searing Wings: Foes that become burned are severely so. The Pokémon's Flying-type moves have a 30% chance to severely burn.