Add new FoN therian abilities
Tornadus: Mending Feather Charm
When the Pokémon is hit thrice, all party members heal for 1/8th HP. The attacker switches out to an ally of your choice.

Thundurus: Raging Scale Charm
When the Pokémon is hit thrice, a Live Wire is set on the foe's side. The attacker hits itself in a rage.

Landorus: Rending Fang Charm
When the Pokémon is hit thrice, two layers of Spikes are set on the foe's side. The attacker will flinch next turn.

Enamorus: Calming Scute Charm
When the Pokémon is hit thrice, a Sanctuary is set on its side for 5 turns. The attacker becomes drowsy.